+++ src/main/java/redstonedubstep/mods/vanishmod/mixin/sound/MixinEntity.java
@@ -9,7 +9,6 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.MoverType;
 import net.minecraft.world.phys.Vec3;
-import redstonedubstep.mods.vanishmod.VanishConfig;
 import redstonedubstep.mods.vanishmod.misc.SoundSuppressionHelper;
 
 @Mixin(Entity.class)
@@ -17,7 +16,7 @@ public class MixinEntity {
 	//Invalidates the hit results of a vanished player if its position changes, because then their crosshair is most likely on a different block
 	@Inject(method = "move", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;setPos(DDD)V", ordinal = 1))
 	private void onActualMove(MoverType type, Vec3 pos, CallbackInfo callbackInfo) {
-		if (VanishConfig.CONFIG.indirectSoundSuppression.get() && (Object)this instanceof ServerPlayer player && !player.hasContainerOpen())
+		if (SoundSuppressionHelper.shouldCapturePlayers() && (Object)this instanceof ServerPlayer player && !player.hasContainerOpen())
 			SoundSuppressionHelper.invalidateHitResults(player);
 	}
 }